requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
azure-core==1.32.0
azure-storage-blob==12.24.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.12.0
colorama==0.4.6
comic2pdf==3.1.0
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==5.1.1
docopt==0.6.2
EbookLib==0.18
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.5
fastavro==1.9.7
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gast==0.6.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.65.1
gspread==6.1.4
gspread-dataframe==4.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.4
img2pdf==0.5.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaleido==0.2.1
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
lida==0.0.14
llmx==0.0.21a0
lxml==5.1.0
m3u8==5.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mdurl==0.1.2
mistune==3.0.2
mizani==0.13.0
ml-dtypes==0.4.0
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.0.8
narwhals==1.15.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
neuralgym @ git+https://github.com/JiahuiYu/neuralgym@88292adb524186693a32404c0cfdc790426ea441
numpy==1.26.4
oauthlib==3.2.2
openai==1.55.3
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
patool==2.2.0
patsy==1.0.1
pickleshare==0.7.5
pikepdf==8.13.0
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
plotnine==0.14.3
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyogrio==0.10.0
pyparsing==3.1.2
pyperclip==1.9.0
pyproj==3.7.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
rarfile==4.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.15.2
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.40.2
sympy==1.12
tenacity==9.0.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.23.1
trio-websocket==0.11.1
typer==0.14.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.32.1
virtualenv==20.28.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wordcloud==1.9.4
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
zipp==3.21.0
